+++ src/components/policy/src/policy/policy_table/table_struct_ext/types.h
@@ -101,7 +101,7 @@ struct ApplicationParams : PolicyBase {
     Optional< AppHMITypes > AppHMIType;
     Optional< RequestTypes > RequestType;
     Optional< Integer<uint16_t, 0, 65225> > memory_kb;
-    Optional< Integer<uint16_t, 0, 65225> > heart_beat_timeout_ms;
+    Optional< Integer<uint32_t, 0, 65225> > heart_beat_timeout_ms;
   public:
     ApplicationParams();
     ApplicationParams(const Strings& groups, Priority priority, HmiLevel default_hmi, bool keep_context, bool steal_focus);
